+++ platform/ios/src/Mapbox.h
@@ -15,6 +15,7 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLCalloutView.h"
 #import "MGLClockDirectionFormatter.h"
 #import "MGLCluster.h"
+#import "MGLCompassButton.h"
 #import "MGLCompassDirectionFormatter.h"
 #import "MGLCoordinateFormatter.h"
 #import "MGLDistanceFormatter.h"